+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.25](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.5.24...@kong-ui-public/analytics-geo-map@1.5.25) (2026-05-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.5.24](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.5.23...@kong-ui-public/analytics-geo-map@1.5.24) (2026-05-10)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
 ## [1.5.23](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.5.22...@kong-ui-public/analytics-geo-map@1.5.23) (2026-05-07)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map