+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.46.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.46.1...@kong-ui-public/entities-shared@3.46.2) (2026-05-12)
+
+
+### Bug Fixes
+
+* **entities-shared:** filter out date fields from code to copy ([#3244](https://github.com/Kong/public-ui-components/issues/3244)) ([e65bc43](https://github.com/Kong/public-ui-components/commit/e65bc432210a38a24237aeeac0cf986fef1c03a4))
+
+
+
+
+
+## [3.46.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.46.0...@kong-ui-public/entities-shared@3.46.1) (2026-05-11)
+
+
+### Bug Fixes
+
+* **entities-shared:** z-index on deck command copy button ([#3242](https://github.com/Kong/public-ui-components/issues/3242)) ([8df058e](https://github.com/Kong/public-ui-components/commit/8df058e6c53b355583abf6226a80b0ab9d568b91))
+
+
+
+
+
+# [3.46.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.2...@kong-ui-public/entities-shared@3.46.0) (2026-05-11)
+
+
+### Features
+
+* **entities-shared:** add a secret advice to the decK customization editor ([#3240](https://github.com/Kong/public-ui-components/issues/3240)) ([ebfd53e](https://github.com/Kong/public-ui-components/commit/ebfd53e1ad7294890011ec54fc766fcdd025437e))
+
+
+
+
+
+## [3.45.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.1...@kong-ui-public/entities-shared@3.45.2) (2026-05-08)
+
+
+### Bug Fixes
+
+* **entities-consumer-groups:** query in exact match ([#3229](https://github.com/Kong/public-ui-components/issues/3229)) ([f9e5e43](https://github.com/Kong/public-ui-components/commit/f9e5e435adaad1af9c3fc91a737f36bece3e0496))
+
+
+
+
+
+## [3.45.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.45.0...@kong-ui-public/entities-shared@3.45.1) (2026-05-08)
+
+
+### Bug Fixes
+
+* **entities-shared:** use unredacted command for decK customization editor ([#3230](https://github.com/Kong/public-ui-components/issues/3230)) ([1f4ff15](https://github.com/Kong/public-ui-components/commit/1f4ff151adc7058690742ea76934a803d84a193b))
+
+
+
+
+
+# [3.45.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.44.4...@kong-ui-public/entities-shared@3.45.0) (2026-05-08)
+
+
+### Features
+
+* **workspace:** allow `workspace` and `editId` to be `null` and workspace-scoped route support ([#3208](https://github.com/Kong/public-ui-components/issues/3208)) ([b1c6611](https://github.com/Kong/public-ui-components/commit/b1c66117d7cbc44c05c1eba6384e365d10efc5ef))
+
+
+
+
+
+## [3.44.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.44.3...@kong-ui-public/entities-shared@3.44.4) (2026-05-07)
+
+
+### Bug Fixes
+
+* **locales/en.json:** fix copies in all the entities ([#3221](https://github.com/Kong/public-ui-components/issues/3221)) ([a3bcf52](https://github.com/Kong/public-ui-components/commit/a3bcf526d9aff14cab8e2f3d9ba7810b5167d974))
+
+
+
+
+
 ## [3.44.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.44.2...@kong-ui-public/entities-shared@3.44.3) (2026-05-05)